+++ app/gui/gui-appmgr.js
@@ -79,7 +79,7 @@ async function init( ) {
         cols: [
           { id: 'Edit', label: "&nbsp;", cellType: "button", width :'5%', icon: 'ui-icon-pencil', 
           method: "GET", URL: 'app/customize', setData: [ { resId : 'CustomizeAddApp' } ] } ,
-          { id: "active",      label: "Enabled",   width: "5%",  cellType: "checkbox" },
+          { id: "enabled",     label: "Enabled",   width: "5%",  cellType: "checkbox" },
           { id: "id",          label: "Id",        width: "20%", cellType: "text" },
           { id: "title",       label: "Title",     width: "15%", cellType: "text" },
           { id: "scope",       label: "Scope",     width: "15%", cellType: "text" },
@@ -102,14 +102,21 @@ async function init( ) {
         description: "Add",
         id: 'CustomizeAddAppForm',
         fieldGroups:[{ columns: [
-          { formFields: [{ id: "appId", label: "Id", type: "text" } ]},
-          { formFields: [{ id: "name",  label: "App Title", type: "text" } ]},
-          { formFields: [{ id: "scope", label: "Scope", type: "select", 
-            optionsResource: { resourceURL: 'scope/options', optionValue: 'id',optionField:'name' } } ]},
-          { formFields: [{ id: "tags", label: "Tags (comma separated)", type: "text" } ]},
-          { formFields: [{ id: "role",  label: "Role",  type: "select",
-            options: addOptions([ "appUser", "admin", 'dev', '-' ]) } ]}
-          ] }],
+          { formFields: [ { id: "appId", label: "Id", type: "text" },
+                          { id: "name",  label: "App Title", type: "text" }
+           ]},
+          { formFields: [
+            { id: "scope", label: "Scope", type: "select", 
+              optionsResource: { resourceURL: 'scope/options', optionValue: 'id',optionField:'name' } },
+            { id: "tags", label: "Tags (comma separated)", type: "text" }
+          ]},
+          { formFields: [
+            { id: "img",  label: "Icon", type: "text" },
+            { id: "role",  label: "Role",  type: "select",
+              options: addOptions([ "appUser", "admin", 'dev', '-' ]) }
+          ]},
+          { formFields: [{ id: "enabled",  label: "Enabled", type: "checkbox" } ]},
+        ] }],
         actions : [ 
           { id: "AddFormBtn", actionName: "Add / Change",
             actionURL: 'app',   update: [{ resId:'CustomizeAppsTbl' }],